% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalNumericRangeInput.R
\name{conditionalNumericRangeInput}
\alias{conditionalNumericRangeInput}
\title{Display a Point or Range Input Box Based on a Trigger Input}
\usage{
conditionalNumericRangeInput(
  id,
  label,
  value,
  value2,
  trigger,
  ns = NS(NULL),
  min = NA,
  max = NA,
  step = NA,
  sep = " to "
)
}
\arguments{
\item{id}{The element ID}

\item{label}{A label for the UI element}

\item{value}{`[numeric(1)]` The initial "point" value}

\item{value2}{`[numeric(2)]` The initial "range" value}

\item{trigger}{`[character(1)]` The variable that triggers the switch between
point and range}

\item{ns}{The session namespace}

\item{min}{The minimum allowable value of the inputs}

\item{max}{The maximum allowable value of the inputs}

\item{step}{The step size}

\item{sep}{The separator between the range input boxes}

\item{style}{Styling to apply to the UI tag}
}
\value{
A UI element
}
\description{
This input will display either a single input box or a range input, depending
on whether it is selected in a trigger variable. It always returns both the
point and the range; use `chooseConditionalRange()` to choose the one
currently displayed in the UI.
}
