% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidscreen-package.R
\docType{package}
\name{covidscreen-package}
\alias{covidscreen}
\alias{covidscreen-package}
\title{covidscreen: COVID-19 Asymtomatic Screening Strategy Assessment}
\description{
Covidscreen is a Shiny app that allows decision-makers to assess the benefits of performing asymptomatic testing in their organizations. It is primarily aimed at healthcare organizations, though it is applicable in a much wider range of settings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jesse-smith/covidscreen}
  \item \url{https://jesse-smith.github.io/covidscreen}
  \item Report bugs at \url{https://github.com/jesse-smith/covidscreen/issues}
}

}
\author{
\strong{Maintainer}: Jesse Smith \email{Jesse.Smith@stjude.org}

Authors:
\itemize{
  \item Tomi Mori \email{Tomi.Mori@stjude.org}
}

}
\keyword{internal}
