% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidtest-package.R
\docType{package}
\name{covidtest-package}
\alias{covidtest}
\alias{covidtest-package}
\title{covidtest: COVID-19 Testing Strategy Assessment}
\description{
Covidtest is a Shiny app that allows decision-makers to assess the benefits of performing asymptomatic testing in their organizations. It is primarily aimed at healthcare organizations, though it is applicable in a much wider range of settings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jesse-smith/covidtest}
  \item Report bugs at \url{https://github.com/jesse-smith/covidtest/issues}
}

}
\author{
\strong{Maintainer}: Jesse Smith \email{jesse.smith@stjude.org}

}
\keyword{internal}
