% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nav.R
\name{cs_nav_pad}
\alias{cs_nav_pad}
\title{Add Padding to a Navbar or Navbar Menu}
\usage{
cs_nav_pad(width = "1rem")
}
\arguments{
\item{width}{The width of the padding element (a `span`)}
}
\value{
A `tags$span()` element with `style = paste("width:", width)`
}
\description{
Add Padding to a Navbar or Navbar Menu
}
