% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{cs_sum}
\alias{cs_sum}
\title{Summarize Joint Distribution from \code{cs_dist()}}
\usage{
cs_sum(dt, vac = NULL, inf = NULL, symp = NULL, test = NULL, detect = NULL)
}
\arguments{
\item{dt}{\code{[data.table]} A distribution from \code{cs_dist()}}

\item{vac, inf, symp, test, detect}{\code{[logical(1)]} Variables values specifying
which sub-group to sum over. If \code{NULL}, will sum over all values.}
}
\value{
\code{[double(1)]} The proportion of the organization with all the specified
  characteristics
}
\description{
\code{cs_sum()} calculates the proportion of the organization
that falls within the specified group. Variables are joined
with \code{&}; not specifiying a variable (or setting it to \code{NULL})
includes all of its values.
}
