% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nav.R
\name{ct_nav_ext_link}
\alias{ct_nav_ext_link}
\title{Add a Hyper-linked Item to a Navbar or Navbar Menu}
\usage{
ct_nav_ext_link(title, href, icon = NULL)
}
\arguments{
\item{title}{The title of the item}

\item{href}{The URL to link to}

\item{icon}{An optional icon for the item}
}
\value{
A `bslib::nav_item`
}
\description{
Add a Hyper-linked Item to a Navbar or Navbar Menu
}
