% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tooltip.R
\name{ct_tooltip}
\alias{ct_tooltip}
\title{Add a tooltip to an HTML Element from `tags`}
\usage{
ct_tooltip(..., tooltip = NULL)
}
\arguments{
\item{...}{Elements to pass to `tags$div()`}

\item{tooltip}{Tooltip text - passed to the `title` of `tags$div()`}
}
\value{
A `div` element
}
\description{
Add a tooltip to an HTML Element from `tags`
}
