% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_info.R
\name{info_modal}
\alias{info_modal}
\alias{ct_info_ui}
\alias{ct_info_server}
\title{Add Info Modal to UI}
\usage{
ct_info_ui(
  id,
  label = NULL,
  icon = shiny::icon("info-circle"),
  container = tags$sup
)

ct_info_server(
  ...,
  title = NULL,
  footer = modalButton("Dismiss"),
  size = c("m", "s", "l", "xl"),
  easyClose = TRUE,
  fade = TRUE
)
}
\arguments{
\item{id}{The input ID}

\item{label}{The text to display with the info link, if any}

\item{icon}{The icon to display on the info link}

\item{container}{A container to wrap the `actionLink`, if any. Defaults to
`tags$sup()`.}

\item{...}{UI elements for the body of the modal dialog box}

\item{title}{An optional title for the dialog}

\item{footer}{UI for the footer. Use `NULL` for no footer.}

\item{size}{One of `"s"` for small, `"m"` for medium, `"l"` for large, or
`"xl"` for extra large}

\item{easyClose}{Can the modal dialog be dismissed by clicking outside the
dialog box? If `FALSE`, the modal dialog must be dismissed by clicking on
a `modalButton()` or from a call to `removeModal()` on the server.}

\item{fade}{If `FALSE`, the modal dialog will have no fade-in animation}
}
\description{
`info_ui()` adds an `actionLink()` to the that brings up the modal in
`info_server()`.
}
\keyword{internal}
