% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_input_panel_server.R
\name{input_panel_server}
\alias{input_panel_server}
\title{Show/Hide and Provide Info on an Input Panel}
\usage{
input_panel_server(
  id,
  toggle_time,
  info,
  info_label = paste0(id, "_info()"),
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{The panel ID (technically the id of the title)}

\item{toggle_time}{The amount of time (in seconds) it takes to show/hide
the panel}

\item{info}{A `tagList`, tag, or text to display when info link is clicked}

\item{info_label}{A label for the info observer; defaults to `"{id}_info()"`}

\item{session}{The shiny `session` object; the default is usually appropriate}
}
\value{
The output of a `bindEvent` call
}
\description{
`input_panel_server()` provides methods to show/hide an input panel by
clicking on the title, as well as get more information by clicking on the
info link. The info link should be named `"{id}_info"`.
}
\keyword{internal}
