% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_insert_args.R
\name{insert_args}
\alias{insert_args}
\title{Update Argument in a reactiveValues dist_args Object}
\usage{
insert_args(
  new_arg,
  dist_args,
  i_nm = c("vac", "inf", "symp", "test", "detect"),
  j_nm = NULL
)
}
\arguments{
\item{new_arg}{The replacement for the argument to be updated}

\item{dist_args}{Existing arguments for `calc_dist()`}

\item{i_nm}{The name of the sub-list to be updated}

\item{j_nm}{The name of the specific vector to be updated, if any}
}
\value{
An updated reactiveValues object
}
\description{
`insert_args()` updates a reactiveValues object containing arguments for
`reactive_dist()` without overwriting values elsewhere. It can either update
an entire sublist or a selected vector within a sub-list.
}
\keyword{internal}
