% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{internal-summary}
\alias{internal-summary}
\alias{detected}
\alias{undetected}
\title{Internal Summary Functions}
\usage{
detected(dt, symp = NULL)

undetected(dt)
}
\arguments{
\item{dt}{`[data.table]` Joint distribution from `calc_dist()`
(or `ct_dist()`)}

\item{symp}{Should proportion be out of all (`NULL`), symptomatic (`TRUE`),
or asymptomatic (`FALSE`) population?}
}
\value{
Summary number
}
\description{
* `detected()` returns the proportion of detected cases
* `undetected()` returns the proportion of undetected cases
}
\keyword{internal}
