% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_profiling_input.R
\name{num_input2}
\alias{num_input2}
\title{Wrap `conditonalNumericRangeInput()` for Ease of Use}
\usage{
num_input2(
  id,
  label,
  value,
  value2,
  min = 0,
  max = NA,
  step = 1,
  trigger = "x_var",
  br = TRUE
)
}
\arguments{
\item{br}{Add a line break below the input element}
}
\description{
Wraps the `conditonalNumericRangeInput()` defined elsewhere in covidtest for
brevity.
}
