% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_dist.R
\name{reactive_dist}
\alias{reactive_dist}
\title{Update the Joint Distribution in Shiny}
\usage{
reactive_dist(dist_args)
}
\arguments{
\item{dist_args}{`[reactiveValues]` A `reactiveValues` object containing
reactives corresponding to the inputs to `calc_dist()`}
}
\value{
A `reactive` containing the updated distribution
}
\description{
Update the Joint Distribution in Shiny
}
