% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalNumericRangeInput.R
\name{reactivePointRange}
\alias{reactivePointRange}
\title{Reactively Select the Point or Range from a Conditional Range Input}
\usage{
reactivePointRange(id, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{The ID of the input}

\item{session}{The Shiny `session`; the default is usually appropriate}
}
\value{
A reactive expression containing the chose input value
}
\description{
Use this function to select the point or range value from a
`conditionalNumericRangeInput()` or a `conditionalSliderRangeInput()`. It
finds the trigger dependency and selects the appropriate value. Unlike
`selectPointRange()`, this returns a reactive expression.
}
