% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_profiling_input.R
\name{select_input}
\alias{select_input}
\title{Select an Input Variable}
\usage{
select_input(id, label = "Input (X-axis)", width = "21rem")
}
\arguments{
\item{id}{The HTML element id}

\item{label}{The UI element label}

\item{width}{the UI element width}
}
\value{
A UI element
}
\description{
Create a dropdown UI element for selecting an input variable to profile over
}
\keyword{internal}
