% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_profiling_input.R
\name{select_output}
\alias{select_output}
\title{Select an Output Type for Plotting}
\usage{
select_output(id)
}
\arguments{
\item{id}{An HTML id for the UI element}
}
\value{
A string
}
\description{
Select an Output Type for Plotting
}
\keyword{internal}
