% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalNumericRangeInput.R
\name{selectPointRange}
\alias{selectPointRange}
\title{Statically elect the Point or Range from a Conditional Range Input}
\usage{
selectPointRange(x, session = getDefaultReactiveDomain())
}
\arguments{
\item{x}{The value from a conditional range input}

\item{session}{The Shiny `session`; the default is usually appropriate}
}
\value{
Either a length 1 ("point") or length 2 ("range") numeric
}
\description{
Use this function to choose the point or range from either a
`conditionalNumericRangeInput()` or `conditionalSliderRangeInput()`. It
finds the trigger dependency and choose the value accordingly.
}
\keyword{internal}
