% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_profiling_input.R
\name{slider_pct2}
\alias{slider_pct2}
\title{Wrap `conditionalSliderRangeInput()` for Ease of Use}
\usage{
slider_pct2(
  id,
  label,
  value = 50,
  value2 = c(0, 100),
  min = 0,
  max = 100,
  step = 1,
  trigger = "x_var",
  prefix = NULL,
  br = TRUE
)
}
\arguments{
\item{id}{The input element ID}

\item{label}{The input element label}

\item{value}{`[numeric(1)]` The point value for the single handle}

\item{value2}{`[numeric(2)]` The range of the double handles}

\item{min}{The minimum slider value}

\item{max}{The maximum slider value}

\item{step}{The step size of the slider}

\item{trigger}{`[character(1)]` The trigger variable}

\item{prefix}{A prefix to prepend to the value labels}

\item{br}{Add a line break below the input element}
}
\value{
A UI element
}
\description{
Wraps the `conditionalSliderRangeInput()` defined elsewhere in covidtest for
brevity.
}
