% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{tabs}
\alias{tabs}
\alias{tab_home}
\alias{tab_scenarios}
\alias{tab_profiling}
\alias{tab_faq}
\alias{nav_menu}
\title{Define Tabs in App}
\usage{
tab_home()

tab_scenarios()

tab_profiling()

tab_faq()

nav_menu()
}
\value{
A `tabPanel` or `navbarMenu`
}
\description{
Define Tabs in App
}
\keyword{internal}
