% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_toggle_panel.R
\name{toggle_panel_server}
\alias{toggle_panel_server}
\title{Handle a toggle panel in the Server}
\usage{
toggle_panel_server(id, input, anim = TRUE, animType = "slide", time = 0.5)
}
\arguments{
\item{id}{The element ID}

\item{input}{The shiny session input}

\item{anim}{If \code{TRUE} then animate the behaviour}

\item{animType}{The type of animation to use, either \code{"slide"} or \code{"fade"}}

\item{time}{The number of seconds to make the animation last}
}
\value{
The return value of `bindEvent()`
}
\description{
A toggle panel is a panel with show/hide functionality by clicking on the
title and additional info by clicking on an info button.
`toggle_panel_server()` defines the logic for show/hide and click events.
}
\keyword{internal}
