% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_dist.R
\name{update_p}
\alias{update_p}
\title{Update the Probabilities for the Joint Distribution}
\usage{
update_p(data, var, dt_p)
}
\arguments{
\item{data}{A `data.table` containing the current distribution, as created
by `calc_dist()`}

\item{var}{A string indicating the variable probability to update}

\item{dt_p}{A `data.table` of probabilities and variable values for joining}
}
\value{
The data with updated probabilities
}
\description{
Update the Probabilities for the Joint Distribution
}
