% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalNumericRangeInput.R
\name{updateConditionalNumericRangeInput}
\alias{updateConditionalNumericRangeInput}
\title{Change the Value or Label of a conditionalNumericRangeInput}
\usage{
updateConditionalNumericRangeInput(
  id,
  label = NULL,
  value = NULL,
  value2 = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{The element ID}

\item{label}{A label for the UI element}

\item{value}{`[numeric(1)]` The initial "point" value}

\item{value2}{`[numeric(2)]` The initial "range" value}

\item{session}{The shiny `session` object; the default is usually appropriate}
}
\value{
The output of `session$sendInputMessage()`
}
\description{
Sends a message to the client to update the either `value`, `value2`,
or `label` in the `conditionalNumericRangeInput()`
}
