% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{cs_dist}
\alias{cs_dist}
\title{Calculate the Steady-State Joint Distribution for a Population}
\usage{
cs_dist(
  vac = list(p_comm = 0.5, p_org = 0.5, eff = 0.5),
  inf = list(p_incid = 0.005, t_symp = 5, t_presymp = 5),
  symp = list(p_inf_vac = 0.5, p_inf_unvac = 0.5, p_uninf = 0),
  test = list(p_symp = 1, p_asymp_vac = 0, p_asymp_unvac = 1/7),
  detect = list(sens = 0.85, spec = 1)
)
}
\arguments{
\item{vac}{`[list(3)]` A named list containing vaccination parameters:
\describe{
  \item{p_comm `[numeric(1)]`}{Proportion vaccinated in the community}
  \item{p_org `[numeric(1)]`}{Proportion vaccinated in the organization of interest}
  \item{eff `[numeric(1)]`}{Vaccine efficacy}
}}

\item{inf}{`[list(3)]` A named list containing infection parameters:
\describe{
  \item{p_incid `[numeric(1)]`}{Proportion of community newly infected each day}
  \item{t_symp `[numeric(1)]`}{Duration of symptomatic period}
  \item{t_presymp `[numeric(1)]`}{Duration of presymptomatic period}
}}

\item{symp}{`[list(3)]` A named lust containing symptom parameters:
\describe{
  \item{p_inf_vac}{Proportion of vaccinated infections who are symptomatic}
  \item{p_inf_unvac}{Proportion of unvaccinated infections who are symptomatic}
  \item{p_uninf}{Proportion of uninfected people who are symptomatic}
}}

\item{test}{`[list(3)]` A named list containing testing parameters:
\describe{
  \item{p_symp `[numeric(1)]`}{Probability of being tested if symptomatic}
  \item{p_asymp_vac `[numeric(1)]`}{Probability of being tested if asymptomatic and vaccinated}
  \item{p_asymp_unvac `[numeric(1)]`}{Probability of being tested if asymptomatic and unvaccinated}
}}

\item{detect}{`[list(2)]` A named list containing detection parameters:
\describe{
  \item{sens `[numeric(1)]`}{Test sensitivity}
  \item{spec `[numeric(1)]`}{Test specificity}
}}
}
\value{
A `data.table`
}
\description{
`cs_dist()` calculates the discrete joint distribution of vaccination,
infection, symptoms, tests, and detections in a population.
}
