% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_profiling_input.R
\name{num_input2}
\alias{num_input2}
\title{Wrap `conditionalNumericRangeInput()` for Ease of Use}
\usage{
num_input2(
  id,
  label,
  value,
  value2,
  min = 0,
  max = NA,
  step = 1,
  trigger = "x_var",
  br = TRUE
)
}
\arguments{
\item{id}{The element ID}

\item{label}{A label for the UI element}

\item{value}{`[numeric(1)]` The initial "point" value}

\item{value2}{`[numeric(2)]` The initial "range" value}

\item{min}{The minimum allowable value of the inputs}

\item{max}{The maximum allowable value of the inputs}

\item{step}{The step size}

\item{trigger}{`[character(1)]` The variable that triggers the switch between
point and range}

\item{br}{Add a line break below the input element}
}
\value{
A UI element
}
\description{
Wraps the `conditionalNumericRangeInput()` defined elsewhere in covidtest for
brevity.
}
\keyword{internal}
