% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{reactive_dist}
\alias{reactive_dist}
\title{Reactive Version of `calc_dist()`}
\usage{
reactive_dist(dist_args, isolate = FALSE)
}
\arguments{
\item{dist_args}{List of arguments to `calc_dist()`}

\item{isolate}{Should the reactive expression be isolated? Will not return
a reactive or take reactive dependencies if `TRUE`.}
}
\value{
A `reactive` containing the results of `calc_dist()`
}
\description{
Takes a `reactiveValues` object and returns a `reactive`
}
\keyword{internal}
