% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_dist.R
\name{reactive_dist}
\alias{reactive_dist}
\title{Update the Joint Distribution in Shiny}
\usage{
reactive_dist(data, vac, inf, symp, test, detect)
}
\arguments{
\item{data}{A `data.table` containing the current distribution, as created
by `calc_dist()`}

\item{vac}{`[list(3)]` A named list containing vaccination parameters:
\describe{
  \item{p_comm `[numeric(1)]`}{Proportion vaccinated in the community}
  \item{p_org `[numeric(1)]`}{Proportion vaccinated in the organization of interest}
  \item{eff `[numeric(1)]`}{Vaccine efficacy}
}}

\item{inf}{`[list(3)]` A named list containing infection parameters:
\describe{
  \item{p_incid `[numeric(1)]`}{Proportion of community newly infected each day}
  \item{t_symp `[numeric(1)]`}{Duration of symptomatic period}
  \item{t_presymp `[numeric(1)]`}{Duration of presymptomatic period}
}}

\item{symp}{`[list(3)]` A named lust containing symptom parameters:
\describe{
  \item{p_inf_vac}{Proportion of vaccinated infections who are symptomatic}
  \item{p_inf_unvac}{Proportion of unvaccinated infections who are symptomatic}
  \item{p_uninf}{Proportion of uninfected people who are symptomatic}
}}

\item{test}{`[list(3)]` A named list containing testing parameters:
\describe{
  \item{p_symp `[numeric(1)]`}{Probability of being tested if symptomatic}
  \item{p_asymp_vac `[numeric(1)]`}{Probability of being tested if asymptomatic and vaccinated}
  \item{p_asymp_unvac `[numeric(1)]`}{Probability of being tested if asymptomatic and unvaccinated}
}}

\item{detect}{`[list(2)]` A named list containing detection parameters:
\describe{
  \item{sens `[numeric(1)]`}{Test sensitivity}
  \item{spec `[numeric(1)]`}{Test specificity}
}}
}
\value{
A `reactive` containing the updated distribution
}
\description{
Update the Joint Distribution in Shiny
}
