% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_dist.R
\name{reactive_dist}
\alias{reactive_dist}
\title{Update the Joint Distribution in Shiny}
\usage{
reactive_dist(dist_args)
}
\arguments{
\item{data}{A `data.table` containing the current distribution, as created
by `calc_dist()`}
}
\value{
A `reactive` containing the updated distribution
}
\description{
Update the Joint Distribution in Shiny
}
