% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_profiling_input.R
\name{reactive_range}
\alias{reactive_range}
\title{Wrap `reactivePointRange()` for Consistent Nomenclature in Main App}
\usage{
reactive_range(id)
}
\arguments{
\item{id}{The ID of the input}
}
\value{
A reactive expression containing the chose input value
}
\description{
Wrap `reactivePointRange()` for Consistent Nomenclature in Main App
}
