% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{risk-metrics}
\alias{risk-metrics}
\alias{cs_undetected}
\alias{cs_rr}
\alias{cs_cost_eff}
\title{Risk-Based Metrics}
\usage{
cs_undetected(dt)

cs_rr(dt, relative = TRUE)

cs_cost_eff(dt)
}
\arguments{
\item{dt}{`[data.table]` The joint distribution from a `cs_dist()`}

\item{relative}{`[logical(1)]` Whether to return risk reduction relative
to baseline risk (`TRUE`) or as an absolute proportion of the organization
(`FALSE`)}
}
\value{
`[double]` A proportion
}
\description{
Undetected cases are the primary measure of risk in the model. From them,
one can calculate the risk reduction and the cost effectiveness of testing.

`cs_undetected()` calculates the proportion of undetected cases
`cs_rr()` calculates the reduction in risk relative to no screening
`cs_cost_eff()` calculates the number of cases detected per test
}
