% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_profiling_input.R
\name{slider_pct2}
\alias{slider_pct2}
\title{Wrap `conditionalSliderRangeInput()` for Ease of Use}
\usage{
slider_pct2(
  id,
  label,
  value = 50,
  value2 = c(0, 100),
  min = 0,
  max = 100,
  step = 1,
  trigger = "x_var",
  prefix = NULL,
  br = TRUE
)
}
\arguments{
\item{br}{Add a line break below the input element}
}
\description{
Wraps the `conditonalSliderRangeInput()` defined elsewhere in covidtest for
brevity.
}
