% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tooltip.R
\name{tag_tooltip}
\alias{tag_tooltip}
\alias{cs_h1}
\alias{cs_h2}
\alias{cs_h3}
\alias{cs_h4}
\alias{cs_h5}
\alias{cs_h6}
\title{HTML Headers with tooltips}
\usage{
cs_h1(..., tooltip = NULL)

cs_h2(..., tooltip = NULL)

cs_h3(..., tooltip = NULL)

cs_h4(..., tooltip = NULL)

cs_h5(..., tooltip = NULL)

cs_h6(..., tooltip = NULL)
}
\arguments{
\item{...}{Inputs to the header function}

\item{tooltip}{Tooltip text}
}
\value{
A `tag` wrapped in a `div` element
}
\description{
Shiny `tags$h()` functions with tooltips
}
